déplacement des fonctions d'affichage dans affichage.py
- déplacement de FPSCLOCK, gameDisplay dans variables.py
- renommage de init_plateau et actualisation_plateau en init_affichage_plateau et actualisation_affichage_plateau
- déplacements des fonctions : affiche_accueil, init_jeu, init_affichage_plateau et actualisation_affichage_plateau, dessine_carte
