+++ variables.py
@@ -22,6 +22,9 @@ GREY=(50,50,50)
 
 espace = 500
 
+FPSCLOCK = pygame.time.Clock()
+gameDisplay = pygame.display.set_mode((WINWIDTH, WINHEIGHT),pygame.RESIZABLE)
+
 #dictionnaire contenant l'ensemble des images du jeu
 IMAGES_DICT = {'titre': pygame.image.load('images/ecran_titre/titre.png'),
 			   'bouton_nouv_jeu':pygame.image.load('images/ecran_titre/bouton_nouv_jeu.png'),
